+++ src/SDL/Input/Mouse.hs
@@ -53,7 +53,7 @@ import Foreign.Ptr
 import Foreign.Storable
 import GHC.Generics (Generic)
 import SDL.Vect
-import SDL.Exception
+import SDL.Internal.Exception
 import SDL.Internal.Numbered
 import SDL.Internal.Types (Window(Window))
 import SDL.Video.Renderer (Surface(Surface))