+++ src/SDL/Video/OpenGL.hs
@@ -37,7 +37,7 @@ import Foreign hiding (void, throwIfNull, throwIfNeg, throwIfNeg_)
 import Foreign.C.Types
 import GHC.Generics (Generic)
 import SDL.Vect
-import SDL.Exception
+import SDL.Internal.Exception
 import SDL.Internal.Numbered
 import SDL.Internal.Types
 import qualified SDL.Raw as Raw