+++ src/SDL/Video/Renderer.hs
@@ -133,7 +133,7 @@ import Foreign.Storable
 import GHC.Generics (Generic)
 import Prelude hiding (foldr)
 import SDL.Vect
-import SDL.Exception
+import SDL.Internal.Exception
 import SDL.Internal.Numbered
 import SDL.Internal.Types
 import qualified Data.ByteString as BS